self CR
