+++ vortex-scalar/src/lib.rs
@@ -23,10 +23,8 @@ mod arrow;
 mod bigint;
 mod binary;
 mod bool;
-mod comparison;
 mod decimal;
 mod display;
-mod error_utils;
 mod extension;
 mod list;
 mod null;
